Fix warnings building on x64. (#486)
* Fix warnings building on x64.

* Remove obsolete compiler flag for x86.

---------

+++ blakserv/osd_windows.c
@@ -39,7 +39,7 @@ void RunMainLoop(void)
 			   case WM_BLAK_MAIN_READ :
 				   EnterServerLock();
 	       
-				   PollSession(msg.lParam);
+				   PollSession((int) msg.lParam);
 				   TimerActivate();
 	       
 				   LeaveServerLock();
@@ -50,13 +50,13 @@ void RunMainLoop(void)
 				   break;
 			   case WM_BLAK_MAIN_DELETE_ACCOUNT :
 				   EnterServerLock();
-				   DeleteAccountAndAssociatedUsersByID(msg.lParam);
+				   DeleteAccountAndAssociatedUsersByID((int) msg.lParam);
 				   LeaveServerLock();
 				   break;
 
 			   case WM_BLAK_MAIN_VERIFIED_LOGIN :
 				   EnterServerLock();
-				   VerifiedLoginSession(msg.lParam);
+				   VerifiedLoginSession((int) msg.lParam);
 				   LeaveServerLock();
 				   break;
